+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 🎁 New Features
 
+* Grid filters configured with `GridFilterFieldSpec.enableValues` offer autocomplete suggestions
+  for 'Equals' and 'Not Equals' filters.
+* `GridFilterFieldSpec` has new `values` and `forceSelection` configs.
 * `FilterChooser` displays a list of fields configured for filtering to improve the usability /
   discoverability of the control. Enabled by default, but can be disabled via
   new `suggestFieldsWhenEmpty` model config.
@@ -18,6 +21,7 @@
 
 ### 💥 Breaking Changes
 
+* `FilterChooserFieldSpec.suggestValues` has been renamed `enableValues`, and now only accepts a boolean.
 * `Store.reuseRecords` must now be explicitly set on Stores with large datasets that wish to cache
   records by raw data identity (see above).
 * `Record` class renamed to `StoreRecord` in anticipation of upcoming changes to JavaScript standard